Fixed stringifying null and undefined
Let's try this again but with a few tests.

In the browser, window.localStorage converts null to "null" and
undefined to "undefined" for both the key and the value. This is
currently inconsistent with the package's behaviour, which is to throw
an error because null and undefined don't have a toString() method.
This should fix the issue.