Changed how keys and values are interpreted. (#53)
* Changed how keys and values are interpreted.

Using x.toString() prevents values like null and undefined from being used as keys or values, which is inconsistent with window.localStorage in the browser. Using "#{x}" instead allows the use of values that do not have .toString() methods, which is consistent with how the Web Storage API functions.

* Fixed stringifying null and undefined

Let's try this again but with a few tests.

In the browser, window.localStorage converts null to "null" and
undefined to "undefined" for both the key and the value. This is
currently inconsistent with the package's behaviour, which is to throw
an error because null and undefined don't have a toString() method.
This should fix the issue.

* Triggered build.

Activated Travis CI. It didn't automatically build the PR so this
should hopefully update the PR so Travis sees it.